feat: Created RoomsTransferFields to refactor ModalBulkTransfer
