+++ src/components/chats/__tests__/ModalBulkTransfer.spec.js
@@ -10,19 +10,6 @@ import i18n from '@/plugins/i18n';
 
 import ModalBulkTransfer from '../chat/ModalBulkTransfer.vue';
 
-jest.mock('@/services/api/resources/settings/queue', () => ({
-  listByProject: jest.fn(() => ({
-    results: [
-      { name: 'Queue 1', sector_name: 'Sector 1', uuid: '1' },
-      { name: 'Queue 2', sector_name: 'Sector 2', uuid: '2' },
-    ],
-  })),
-  agentsToTransfer: jest.fn(() => [
-    { first_name: 'John', last_name: 'Doe', email: 'john@doe.com' },
-    { name: 'Jane', sector_name: 'Doe', uuid: 'jane@doe.com' },
-  ]),
-}));
-
 const localVue = createLocalVue();
 
 localVue.use(Vuex);
@@ -83,51 +70,8 @@ describe('ModalBulkTransfer', () => {
     });
   });
 
-  describe('Data Loading', () => {
-    it('should load queues when created', async () => {
-      expect(wrapper.vm.queues.length).toBeGreaterThan(1);
-    });
-  });
-
-  describe('Field Behavior', () => {
-    it('should disable agent field when queue is not selected or do not have agents to select', async () => {
-      const agentSelect = wrapper.find('[data-testid="select-agent"]');
-
-      wrapper.setData({
-        selectedQueue: [{ value: 'queue_id', label: 'Queue' }],
-        agents: [
-          { value: '', label: 'Select agent' },
-          { value: 'agent2_id', label: 'Agent2' },
-        ],
-      });
-      await wrapper.vm.$nextTick();
-      expect(agentSelect.props('disabled')).toBe(false);
-
-      wrapper.setData({
-        selectedQueue: [{ value: '', label: 'Select queue' }],
-        agents: [
-          { value: '', label: 'Select agent' },
-          { value: 'agent2_id', label: 'Agent2' },
-        ],
-      });
-      await wrapper.vm.$nextTick();
-      expect(agentSelect.props('disabled')).toBe(true);
-
-      wrapper.setData({
-        selectedQueue: [{ value: 'queue_id', label: 'Queue' }],
-        agents: [{ value: '', label: 'Select agent' }],
-      });
-      await wrapper.vm.$nextTick();
-      expect(agentSelect.props('disabled')).toBe(false);
-    });
-  });
-
   describe('Bulk Transfer', () => {
     it('should perform bulk transfer when transfer button is clicked', async () => {});
-
-    it('should show success alert after successful bulk transfer', async () => {});
-
-    it('should show error alert if bulk transfer fails', async () => {});
 
     it('should close modal after successful bulk transfer', async () => {});
   });
@@ -141,12 +85,6 @@ describe('ModalBulkTransfer', () => {
   describe('Modal Interaction', () => {
     it('should close modal after clicking cancel button', () => {});
     it('should emit close event when modal is closed', () => {});
-  });
-
-  describe('Alerts', () => {
-    it('should display success alert with correct message', () => {});
-
-    it('should display error alert with correct message', () => {});
   });
 
   describe('Localization', () => {
